Light Sensor
Included is the .c and .h files for the light sensor. Adjust the second
period to tune the period for the lightsense. There are 2 issues we
should address. First I need some sort of flag of where the robot is
travelling, stafing left, right, moving forward, backward, ect. Also
light calibration for the light sensors. I think I can write something
in the higher level machine code, but that would be up to the raspi, as
it is the master. (Swiping across the line with motors, measuring how
long the decay takes to reach 0, on black and white, then update the
value in the program to adjust it)